over

+++ src/main/java/com/yangliuxin/controller/wechat/WeChatController.java
@@ -423,7 +423,7 @@ public class WeChatController {
         ResultVo<Shop> resultVo = new ResultVo<>();
         String today = LocalDate.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"));
         Shop shop = shopRepository.getShopData(shopId, today);
-        shop.setLotteryCount((Long)redisTemplate.opsForValue().get("GIFT_SHOP_ID_DAY_"+LocalDate.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"))));
+        shop.setLotteryCount((Integer)redisTemplate.opsForValue().get("GIFT_SHOP_ID_DAY_"+LocalDate.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"))));
         resultVo.setCode(1);
         resultVo.setMsg("请求成功");
         resultVo.setData(shop);
